# test-constructor/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend part of test constructor",
        "title": "test constructor",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/manager/events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список мероприятий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Получить мероприятия",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/manager.Event"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Создать настройку мероприятия",
                "parameters": [
                    {
                        "description": "EventCfg object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/manager.EventCfgInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/manager/events/{id}/specializations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение списка специализаций конкретного мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Получить специализации мероприятия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/manager.EventSpecializationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/manager/tests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список тестов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Получить тесты для организотора",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/manager.TestsInfoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать новый тест",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Создать тест",
                "parameters": [
                    {
                        "description": "Test object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/manager.CreateTestInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/manager/tests/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Удалить тест",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "manager.CreateQuestionInfo": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/models.QuestionOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "manager.CreateTestInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.CreateQuestionInfo"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "manager.Event": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "manager.EventCfgInfo": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "extra_threshold": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.ExtraThresholdInfo"
                    }
                },
                "fail_text": {
                    "type": "string"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "success_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
        "manager.EventSpecializationsResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_name": {
                    "type": "string"
                },
                "specializations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.Specialization"
                    }
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "manager.ExtraThresholdInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
        "manager.Specialization": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "manager.TestInfo": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "manager.TestsInfoResponse": {
            "type": "object",
            "properties": {
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.TestInfo"
                    }
                }
            }
        },
        "models.ChoiceOption": {
            "type": "object",
            "properties": {
                "is_true": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.MatchingPair": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "string"
                },
                "right": {
                    "type": "string"
                }
            }
        },
...
        "models.QuestionOptions": {
            "type": "object",
            "properties": {
                "case_sensitive": {
                    "type": "boolean"
                },
                "choice": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChoiceOption"
                    }
                },
                "correct_input": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "matching": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MatchingPair"
                    }
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SequenceItem"
                    }
                }
            }
        },
        "models.SequenceItem": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}